% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey_data.R
\name{get_survey_data}
\alias{get_survey_data}
\title{Get the survey data from a survey with the survey ID}
\usage{
get_survey_data(survey_id = Sys.getenv("QUALTRICS_SURVEY_ID"))
}
\arguments{
\item{survey_id}{Survey ID. By default, it uses \code{QUALTRICS_SURVEY_ID} in the environment variable.}
}
\value{
A tibble of survey responses
}
\description{
Get the survey data from a survey specified by the survey ID. The function requires that the
}
\examples{
responses <- get_survey_data()
}
