% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navbar.R
\name{navbar_UI}
\alias{navbar_UI}
\title{Title Render the navbar UI}
\usage{
navbar_UI(id, logo)
}
\arguments{
\item{id}{id}

\item{logo}{a string specifying the location of an image file}
}
\value{
a Shiny \code{tagList} object
}
\description{
Title Render the navbar UI
}
\examples{
navbar_UI("navbar", logo = "https://ui-avatars.com/api/?name=PostReview")
}
