% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_promoter_score.R
\name{net_promoter_score_server}
\alias{net_promoter_score_server}
\title{Net promoter score server module}
\usage{
net_promoter_score_server(id, data)
}
\arguments{
\item{id}{element ID}

\item{data}{survey responses data}
}
\value{
shiny output
}
\description{
Net promoter score server module
}
\examples{
net_promoter_score_server("nps")
}
