% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_promoter_score.R
\name{net_promoter_score_UI}
\alias{net_promoter_score_UI}
\title{Net promoter score UI module}
\usage{
net_promoter_score_UI(id)
}
\arguments{
\item{id}{element ID}
}
\value{
a \code{div} tag object
}
\description{
Net promoter score UI module
}
\examples{
net_promoter_score_UI("nps")
}
