% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_app}
\alias{run_app}
\title{Run The WRK Survey App}
\usage{
run_app(...)
}
\arguments{
\item{...}{parameters passed to \code{shiny::shinyApp()}}
}
\value{
an app
}
\description{
Run The WRK Survey App
}
\examples{
\dontrun{
run_app()
}
}
